limbo: move cyclical cases to `pathological` namespace
