+++ limbo/testcases/__init__.py
@@ -5,6 +5,7 @@ Models and definitions for generating Limbo testcases.
 from limbo.testcases._core import registry
 
 from .pathlen import *  # noqa: F403
+from .pathological import *  # noqa: F403
 from .rfc5280 import *  # noqa: F403
 from .webpki import *  # noqa: F403
